+++ src/beamlime/kafka/consumer.py
@@ -10,7 +10,7 @@ from confluent_kafka.error import KafkaException
 
 from .. import StreamKind
 from ..config.config_loader import load_config
-from ..config.stream_mapping import stream_kind_to_topic
+from ..config.streams import stream_kind_to_topic
 
 
 def validate_topics_exist(consumer: kafka.Consumer, topics: list[str]) -> None:
