+++ src/beamlime/kafka/routes.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2025 Scipp contributors (https://github.com/scipp)
 
-from ..config.stream_mapping import StreamMapping, stream_kind_to_topic
+from ..config.streams import StreamMapping, stream_kind_to_topic
 from ..core.message import StreamKind
 from .message_adapter import (
     BeamlimeConfigMessageAdapter,
